+++ packages/backend/src/analytics/LightdashAnalytics.ts
@@ -202,16 +202,24 @@ type TrackUserDeletedEvent = BaseTrack & {
     };
 };
 
-type TrackSavedChart = BaseTrack & {
-    event: 'saved_chart.updated' | 'saved_chart.deleted';
+type UpdateSavedChartEvent = BaseTrack & {
+    event: 'saved_chart.updated';
+    properties: {
+        projectId: string;
+        savedQueryId: string;
+        dashboardId: string | undefined;
+    };
+};
+type DeleteSavedChartEvent = BaseTrack & {
+    event: 'saved_chart.deleted';
     properties: {
         projectId: string;
         savedQueryId: string;
     };
 };
 
-export type CreateSavedChartOrVersionEvent = BaseTrack & {
-    event: 'saved_chart.created' | 'saved_chart_version.created';
+export type CreateSavedChartVersionEvent = BaseTrack & {
+    event: 'saved_chart_version.created';
     properties: {
         projectId: string;
         savedQueryId: string;
@@ -243,6 +251,13 @@ export type CreateSavedChartOrVersionEvent = BaseTrack & {
         bigValue?: {
             hasBigValueComparison?: boolean;
         };
+    };
+};
+
+export type CreateSavedChartEvent = BaseTrack & {
+    event: 'saved_chart.created';
+    properties: CreateSavedChartVersionEvent['properties'] & {
+        dashboardId: string | undefined;
         duplicated?: boolean;
     };
 };
@@ -332,7 +347,7 @@ type ProjectCompiledEvent = BaseTrack & {
         roundCount?: number;
         formattedFieldsCount?: number;
         urlsCount?: number;
-        rowAccessFiltersCount: number;
+        modelsWithSqlFiltersCount: number;
     };
 };
 
@@ -711,8 +726,10 @@ type Track =
     | VerifiedUserEvent
     | UserJoinOrganizationEvent
     | QueryExecutionEvent
-    | TrackSavedChart
-    | CreateSavedChartOrVersionEvent
+    | UpdateSavedChartEvent
+    | DeleteSavedChartEvent
+    | CreateSavedChartEvent
+    | CreateSavedChartVersionEvent
     | TrackUserDeletedEvent
     | ProjectErrorEvent
     | ApiErrorEvent