+++ packages/backend/src/models/DashboardModel/DashboardModel.ts
@@ -10,6 +10,7 @@ import {
     DashboardUnversionedFields,
     DashboardVersionedFields,
     NotFoundError,
+    SavedChart,
     SessionUser,
     UnexpectedServerError,
     UpdateMultipleDashboards,
@@ -485,6 +486,8 @@ export class DashboardModel {
                     title: string | null;
                     views: string;
                     first_viewed_at: Date | null;
+                    belongs_to_dashboard: boolean;
+                    name: string | null;
                 }[]
             >(
                 `${DashboardTilesTableName}.x_offset`,
@@ -494,6 +497,10 @@ export class DashboardModel {
                 `${DashboardTilesTableName}.height`,
                 `${DashboardTilesTableName}.dashboard_tile_uuid`,
                 `${SavedChartsTableName}.saved_query_uuid`,
+                `${SavedChartsTableName}.name`,
+                this.database.raw(
+                    `${SavedChartsTableName}.dashboard_uuid IS NOT NULL AS belongs_to_dashboard`,
+                ),
                 this.database.raw(
                     `COALESCE(
                         ${DashboardTileChartTableName}.title,
@@ -578,6 +585,8 @@ export class DashboardModel {
                     hide_title,
                     url,
                     content,
+                    belongs_to_dashboard,
+                    name,
                 }) => {
                     const base: Omit<
                         Dashboard['tiles'][number],
@@ -603,6 +612,8 @@ export class DashboardModel {
                                 properties: {
                                     ...commonProperties,
                                     savedChartUuid: saved_query_uuid,
+                                    belongsToDashboard: belongs_to_dashboard,
+                                    chartName: name,
                                 },
                             };
                         case DashboardTileTypes.MARKDOWN:
@@ -772,5 +783,35 @@ export class DashboardModel {
             );
         });
         return this.getById(dashboardUuid);
+    }
+
+    async getOrphanedCharts(
+        dashboardUuid: string,
+    ): Promise<Pick<SavedChart, 'uuid'>[]> {
+        const getLastVersionIdQuery = this.database(DashboardsTableName)
+            .leftJoin(
+                DashboardVersionsTableName,
+                `${DashboardsTableName}.dashboard_id`,
+                `${DashboardVersionsTableName}.dashboard_id`,
+            )
+            .select([`${DashboardVersionsTableName}.dashboard_version_id`])
+            .where(`${DashboardsTableName}.dashboard_uuid`, dashboardUuid)
+            .orderBy(`${DashboardVersionsTableName}.created_at`, 'desc')
+            .limit(1);
+
+        const getChartsInTilesQuery = this.database(DashboardTileChartTableName)
+            .select(`saved_chart_id`)
+            .where(
+                `${DashboardTileChartTableName}.dashboard_version_id`,
+                getLastVersionIdQuery,
+            );
+        const orphanedCharts = await this.database(SavedChartsTableName)
+            .select(`saved_query_uuid`)
+            .where(`${SavedChartsTableName}.dashboard_uuid`, dashboardUuid)
+            .whereNotIn(`saved_query_id`, getChartsInTilesQuery);
+
+        return orphanedCharts.map((chart) => ({
+            uuid: chart.saved_query_uuid,
+        }));
     }
 }