+++ packages/backend/src/services/DashboardService/DashboardService.ts
@@ -7,6 +7,7 @@ import {
     DashboardTileTypes,
     ForbiddenError,
     isChartScheduler,
+    isChartTile,
     isDashboardUnversionedFields,
     isDashboardVersionedFields,
     isSlackTarget,
@@ -27,8 +28,10 @@ import { getFirstAccessibleSpace } from '../../database/entities/spaces';
 import { AnalyticsModel } from '../../models/AnalyticsModel';
 import { DashboardModel } from '../../models/DashboardModel/DashboardModel';
 import { PinnedListModel } from '../../models/PinnedListModel';
+import { SavedChartModel } from '../../models/SavedChartModel';
 import { SchedulerModel } from '../../models/SchedulerModel';
 import { SpaceModel } from '../../models/SpaceModel';
+import { SavedChartService } from '../SavedChartsService/SavedChartService';
 import { hasSpaceAccess } from '../SpaceService/SpaceService';
 
 type Dependencies = {
@@ -37,6 +40,7 @@ type Dependencies = {
     analyticsModel: AnalyticsModel;
     pinnedListModel: PinnedListModel;
     schedulerModel: SchedulerModel;
+    savedChartModel: SavedChartModel;
 };
 
 export class DashboardService {
@@ -50,18 +54,22 @@ export class DashboardService {
 
     schedulerModel: SchedulerModel;
 
+    savedChartModel: SavedChartModel;
+
     constructor({
         dashboardModel,
         spaceModel,
         analyticsModel,
         pinnedListModel,
         schedulerModel,
+        savedChartModel,
     }: Dependencies) {
         this.dashboardModel = dashboardModel;
         this.spaceModel = spaceModel;
         this.analyticsModel = analyticsModel;
         this.pinnedListModel = pinnedListModel;
         this.schedulerModel = schedulerModel;
+        this.savedChartModel = savedChartModel;
     }
 
     static getCreateEventProperties(
@@ -85,6 +93,30 @@ export class DashboardService {
                 ({ type }) => type === DashboardTileTypes.LOOM,
             ).length,
         };
+    }
+
+    private async deleteOrphanedChartsInDashboards(
+        user: SessionUser,
+        dashboardUuid: string,
+    ) {
+        const orphanedCharts = await this.dashboardModel.getOrphanedCharts(
+            dashboardUuid,
+        );
+        await Promise.all(
+            orphanedCharts.map(async (chart) => {
+                const deletedChart = await this.savedChartModel.delete(
+                    chart.uuid,
+                );
+                analytics.track({
+                    event: 'saved_chart.deleted',
+                    userId: user.userUuid,
+                    properties: {
+                        savedQueryId: deletedChart.uuid,
+                        projectId: deletedChart.projectUuid,
+                    },
+                });
+            }),
+        );
     }
 
     async hasDashboardSpaceAccess(
@@ -169,6 +201,40 @@ export class DashboardService {
         return dashboard;
     }
 
+    static findChartsThatBelongToDashboard(
+        dashboard: Pick<Dashboard, 'tiles'>,
+    ): string[] {
+        return dashboard.tiles.reduce<string[]>((acc, tile) => {
+            if (
+                isChartTile(tile) &&
+                !!tile.properties.belongsToDashboard &&
+                !!tile.properties.savedChartUuid
+            ) {
+                return [...acc, tile.properties.savedChartUuid];
+            }
+            return acc;
+        }, []);
+    }
+
+    private async trackChartsCreatedInDashboard(
+        user: SessionUser,
+        chartUuids: string[],
+    ) {
+        await Promise.all(
+            chartUuids.map(async (chartUuid) => {
+                const chart = await this.savedChartModel.get(chartUuid);
+                analytics.track({
+                    event: 'saved_chart.created',
+                    userId: user.userUuid,
+                    properties: {
+                        ...SavedChartService.getCreateEventProperties(chart),
+                        dashboardId: chart.dashboardUuid ?? undefined,
+                    },
+                });
+            }),
+        );
+    }
+
     async create(
         user: SessionUser,
         projectUuid: string,
@@ -216,6 +282,15 @@ export class DashboardService {
             userId: user.userUuid,
             properties: DashboardService.getCreateEventProperties(newDashboard),
         });
+
+        // Track charts created in dashboard
+        const chartsCreatedInDashboard =
+            DashboardService.findChartsThatBelongToDashboard(newDashboard);
+        await this.trackChartsCreatedInDashboard(
+            user,
+            chartsCreatedInDashboard,
+        );
+
         return this.dashboardModel.getById(newDashboard.uuid);
     }
 
@@ -344,6 +419,25 @@ export class DashboardService {
                 properties:
                     DashboardService.getCreateEventProperties(updatedDashboard),
             });
+            await this.deleteOrphanedChartsInDashboards(user, dashboardUuid);
+
+            // Track charts created in dashboard
+            const chartsInExistingDashboard =
+                DashboardService.findChartsThatBelongToDashboard(
+                    existingDashboard,
+                );
+            const chartsInUpdatedDashboard =
+                DashboardService.findChartsThatBelongToDashboard(
+                    updatedDashboard,
+                );
+            const chartsCreatedInDashboard = chartsInUpdatedDashboard.filter(
+                (chart) => !chartsInExistingDashboard.includes(chart),
+            );
+
+            await this.trackChartsCreatedInDashboard(
+                user,
+                chartsCreatedInDashboard,
+            );
         }
         return this.dashboardModel.getById(dashboardUuid);
     }