+++ packages/backend/src/services/services.ts
@@ -84,6 +84,7 @@ export const projectService = new ProjectService({
     emailClient,
     spaceModel,
     sshKeyPairModel,
+    userAttributesModel,
 });
 
 export const shareService = new ShareService({
@@ -102,6 +103,7 @@ export const dashboardService = new DashboardService({
     analyticsModel,
     pinnedListModel,
     schedulerModel,
+    savedChartModel,
 });
 
 export const savedChartsService = new SavedChartService({