+++ packages/common/src/types/savedCharts.ts
@@ -274,6 +274,7 @@ export type SavedChart = {
     pinnedListUuid: string | null;
     pinnedListOrder: number | null;
     dashboardUuid: string | null;
+    dashboardName: string | null;
 };
 
 export type CreateSavedChart = Omit<
@@ -289,6 +290,7 @@ export type CreateSavedChart = Omit<
     | 'views'
     | 'firstViewedAt'
     | 'dashboardUuid'
+    | 'dashboardName'
 > & { spaceUuid?: string };
 
 export type CreateSavedChartVersion = Omit<
@@ -305,6 +307,7 @@ export type CreateSavedChartVersion = Omit<
     | 'views'
     | 'firstViewedAt'
     | 'dashboardUuid'
+    | 'dashboardName'
 >;
 
 export type UpdateSavedChart = Partial<