+++ packages/frontend/src/pages/Dashboard.tsx
@@ -19,6 +19,7 @@ import { useHistory, useParams } from 'react-router-dom';
 
 import DashboardHeader from '../components/common/Dashboard/DashboardHeader';
 import ErrorState from '../components/common/ErrorState';
+import DashboardDeleteModal from '../components/common/modal/DashboardDeleteModal';
 import Page from '../components/common/Page/Page';
 import DashboardFilter from '../components/DashboardFilter';
 import ChartTile from '../components/DashboardTiles/DashboardChartTile';
@@ -31,7 +32,6 @@ import MetricQueryDataProvider from '../components/MetricQueryData/MetricQueryDa
 import UnderlyingDataModal from '../components/MetricQueryData/UnderlyingDataModal';
 import {
     appendNewTilesToBottom,
-    useDashboardDeleteMutation,
     useDuplicateDashboardMutation,
     useExportDashboard,
     useMoveDashboardMutation,
@@ -170,8 +170,9 @@ const Dashboard: FC = () => {
     const { mutate: duplicateDashboard } = useDuplicateDashboardMutation({
         showRedirectButton: true,
     });
-    const { mutateAsync: deleteDashboard } = useDashboardDeleteMutation();
     const { mutate: exportDashboard } = useExportDashboard();
+
+    const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
 
     const layouts = useMemo(
         () => ({
@@ -182,11 +183,27 @@ const Dashboard: FC = () => {
         [dashboardTiles, isEditMode],
     );
 
+    const { tiles: savedTiles } = dashboard || {};
     useEffect(() => {
-        if (dashboard?.tiles) {
-            setDashboardTiles(dashboard.tiles);
+        if (savedTiles) {
+            const unsavedDashboardTilesRaw = sessionStorage.getItem(
+                'unsavedDashboardTiles',
+            );
+            sessionStorage.removeItem('unsavedDashboardTiles');
+            let unsavedDashboardTiles = undefined;
+            if (unsavedDashboardTilesRaw) {
+                try {
+                    unsavedDashboardTiles = JSON.parse(
+                        unsavedDashboardTilesRaw,
+                    );
+                } catch {
+                    // do nothing
+                }
+            }
+            setDashboardTiles(unsavedDashboardTiles || savedTiles);
+            setHaveTilesChanged(!!unsavedDashboardTiles);
         }
-    }, [dashboard, setDashboardTiles]);
+    }, [setHaveTilesChanged, setDashboardTiles, savedTiles]);
 
     useEffect(() => {
         if (isSuccess) {
@@ -275,6 +292,7 @@ const Dashboard: FC = () => {
     );
 
     const handleCancel = useCallback(() => {
+        sessionStorage.clear();
         setDashboardTiles(dashboard?.tiles || []);
         setHaveTilesChanged(false);
         if (dashboard) setDashboardFilters(dashboard.filters);
@@ -310,9 +328,7 @@ const Dashboard: FC = () => {
 
     const handleDeleteDashboard = () => {
         if (!dashboard) return;
-        deleteDashboard(dashboard.uuid).then(() => {
-            history.replace(`/projects/${projectUuid}/dashboards`);
-        });
+        setIsDeleteModalOpen(true);
     };
 
     const handleExportDashboard = () => {
@@ -339,17 +355,21 @@ const Dashboard: FC = () => {
     }, [haveTilesChanged, haveFiltersChanged, isEditMode]);
 
     useEffect(() => {
+        const createChartInDashboardFlow = sessionStorage.getItem(
+            'unsavedDashboardTiles',
+        );
         history.block((prompt) => {
             if (
                 isEditMode &&
                 (haveTilesChanged || haveFiltersChanged) &&
                 !prompt.pathname.includes(
                     `/projects/${projectUuid}/dashboards/${dashboardUuid}`,
-                )
+                ) &&
+                createChartInDashboardFlow
             ) {
                 setBlockedNavigationLocation(prompt.pathname);
                 setIsSaveWarningModalOpen(true);
-                return false; //blocks history
+                return false; // blocks history
             }
             return undefined; // allow history
         });
@@ -480,11 +500,23 @@ const Dashboard: FC = () => {
                         isEditMode={isEditMode}
                     />
                 )}
+
+                {isDeleteModalOpen && (
+                    <DashboardDeleteModal
+                        opened
+                        uuid={dashboard.uuid}
+                        onClose={() => setIsDeleteModalOpen(false)}
+                        onConfirm={() => {
+                            history.replace(
+                                `/projects/${projectUuid}/dashboards`,
+                            );
+                        }}
+                    />
+                )}
             </Page>
         </>
     );
 };
-
 const DashboardPage: FC = () => {
     useProfiler('Dashboard');
     return (